requirements.txt
psycopg2-binary==2.9.8
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
Pillow==10.0.0